requirements.txt
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
#
Flask
python-dateutil
Flask-WTF
markdown
git+https://github.com/r0wb0t/markdown-urlize
flask_sqlalchemy
pytest
PyMySQL
requests
GoogleAppEngineCloudStorageClient